editoast: authz: move `trait TestClientExt` in its own module
It's an aggregation of shortcuts of protected operations from
different resources. Prevents polluting v2.rs as the trait implementation
inevitably grows.

No code changes.
